+++ app/src/main/java/com/example/tidy/ui/component/pickers/TimePickerTidy.kt
@@ -1,0 +1,55 @@
+/*
+ * Copyright (C) 2026  Gaurav Kumar
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package com.example.tidy.ui.component.pickers
+
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.TimePicker
+import androidx.compose.material3.TimePickerState
+import androidx.compose.material3.rememberTimePickerState
+import androidx.compose.runtime.Composable
+import com.example.tidy.ui.component.dialog.SimpleDialog
+import java.util.Calendar
+
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TimePickerTidy(
+    hour: Int?,
+    minute: Int?,
+    onTimeSelected: (TimePickerState) -> Unit,
+    onDismiss: () -> Unit
+) {
+    val currentTime = Calendar.getInstance()
+    val timePickerState = rememberTimePickerState(
+        initialHour = hour ?: currentTime.get(Calendar.HOUR_OF_DAY),
+        initialMinute = minute ?: currentTime.get(Calendar.MINUTE),
+        is24Hour = false,
+    )
+    SimpleDialog(
+        onDismissRequest = onDismiss,
+        onConfirm = {
+            onTimeSelected(timePickerState)
+            onDismiss()
+        },
+        content = {
+            TimePicker(
+                state = timePickerState
+            )
+        },
+        title = "Select Time"
+    )
+}
